#9 => 변경일시 타입 오류 수정

+++ Ami/src/main/java/com/cnu/ami/device/equipment/dao/entity/MeterInfoInterfaceVO.java
@@ -26,6 +26,6 @@ public interface MeterInfoInterfaceVO {
 	
 	int getMrd();
 	
-	int getUDate();
+	long getUDate();
 
 }
